Allow slicing and put LU decomp in square matrix class
